LASolver: Ensure correct models in theory combination
In theory combination, interface variables which are not explicitly set
to be equal are implicitly assumed to be disequal (by Egraph).
This needs to be honored by the LASolver when computing the model.

However, previously we could have a variable with value 0 + delta, which
is considered different than 1. But when we computed actual real values,
delta was computed as 1 (which is OK for the LA model), but this is no
longer OK with the Egraph model, which assumed the variable to be not
equal to 1.

To remedy this situation, we add additional check when figuring out
which interface equalities must be decided.